+++ Client/Scripts/Gameplay/Dto/AbilityContext.cs
@@ -12,7 +12,7 @@ public sealed record AbilityContext : IServiceProvider
     /// <summary>
     /// Snapshot of user stats at the moment the ability took effect. Use this instead of User.Stats.
     /// </summary>
-    public required IReadOnlyDictionary<Stat, float> UserStats { get; init; }
+    public required IReadOnlyDictionary<Stat, double> UserStats { get; init; }
     public required IServiceProvider ServiceProvider { get; init; }
     public Vector2? TargetPosition { get; init; }
     public Unit? TargetUnit { get; init; }
